sync with v8.0.0

+++ Source/GB_BinaryOp_check.c
@@ -68,7 +68,6 @@ GrB_Info GB_BinaryOp_check  // check a GraphBLAS binary operator
     }
     else if (opcode == GB_FIRST_binop_code && op->ztype->code == GB_UDT_code)
     { 
-// GB_GOTCHA ; // FIRST, GB_reduce_to_vector
         // FIRST_UDT binary operator created by GB_reduce_to_vector
         GBPR0 ("(generated): z=%s(x,y)\n", op->name) ;
     }
@@ -91,7 +90,7 @@ GrB_Info GB_BinaryOp_check  // check a GraphBLAS binary operator
     }
 
     int32_t name_len = op->name_len ;
-    int32_t actual_len = strlen (op->name) ;
+    int32_t actual_len = (int32_t) strlen (op->name) ;
     if (opcode == GB_USER_binop_code && name_len != actual_len)
     { 
         GBPR0 ("    BinaryOp has an invalid name_len\n") ;
