+++ Source/GB_Type_check.c
@@ -71,7 +71,7 @@ GrB_Info GB_Type_check      // check a GraphBLAS Type
     }
 
     int32_t name_len = type->name_len ;
-    int32_t actual_len = strlen (type->name) ;
+    int32_t actual_len = (int32_t) strlen (type->name) ;
     if (type->code == GB_UDT_code && name_len != actual_len)
     { 
         GBPR0 ("    Type has an invalid name_len\n") ;
